reafactor: move Username component
