+++ src/components/common/atoms/index.tsx
@@ -4,3 +4,4 @@ export {ContextButton} from './ContextButton';
 export {CopyIDButton} from './CopyIDButton';
 export {Link} from './Link';
 export {Text} from './Text';
+export {Username} from './Username';
