+++ src/components/sheets/ProfileSheet.tsx
@@ -13,9 +13,9 @@ import {User, Server} from 'revolt.js';
 import {app, client, GeneralAvatar, openUrl} from '../../Generic';
 import {BADGES, USER_IDS} from '../../lib/consts';
 import {parseRevoltNodes} from '../../lib/utils';
-import {Avatar, MiniProfile, RoleView, Username} from '../../Profile';
+import {Avatar, MiniProfile, RoleView} from '../../Profile';
 import {currentTheme, styles} from '../../Theme';
-import {Button, ContextButton, Link, Text} from '../common/atoms';
+import {Button, ContextButton, Link, Text, Username} from '../common/atoms';
 import {MarkdownView} from '../common/MarkdownView';
 import {UserMenuSheet} from './index';
 import {UserList} from '../navigation/UserList';
